+++ NEWS.md
@@ -21,6 +21,7 @@
 * Use GitHub Workflows for Maven (APPPOCTOOL-81)
 * Bump Spring Boot from 4.0.3 to 4.0.4 fixing Jackson vulns (APPPOCTOOL-83)
 * Upgrade Keycloak testcontainers to 26.5.7 (KEYCLOAK-102)
+* Support tenant-aware Kafka message filtering (APPPOCTOOL-85)
 
 ### Migration:
 * Replace KeycloakSecretUtils with SecureStoreKeyProvider, add `application.secret-store.environment=${SECURE_STORE_ENV:folio}` to application.yaml (for Ramsons and Sunflower: add `application.secret-store.environment=${SECURE_STORE_ENV:${ENV:folio}}` to application.yaml).